refactor(gc): move store::fs::gc -> store::gc
