+++ src/store/fs/options.rs
@@ -4,8 +4,8 @@ use std::{
     time::Duration,
 };
 
-pub use super::gc::{GcConfig, ProtectCb, ProtectOutcome};
 use super::{meta::raw_outboard_size, temp_name};
+pub use crate::store::gc::{GcConfig, ProtectCb, ProtectOutcome};
 use crate::Hash;
 
 /// Options for directories used by the file store.
